feat: separate login modal logic
