+++ pages/index.tsx
@@ -6,7 +6,6 @@ const Index: NextPage = () => {
     const router = useRouter();
 
     useEffect(() => {
-        // eslint-disable-next-line @typescript-eslint/no-floating-promises
         router.push('/maps/mine');
     }, []);
